refactor(console): app containers
