+++ packages/console/src/containers/AppContent/index.module.scss
@@ -1,0 +1,11 @@
+@use '@/scss/underscore' as _;
+
+.main {
+  flex-grow: 1;
+  padding: 0 _.unit(2);
+  overflow-y: scroll;
+
+  > * {
+    @include _.main-content-width;
+  }
+}
