+++ packages/console/src/containers/AppLayout/components/Sidebar/components/Item/index.module.scss
