+++ packages/console/src/containers/AppLayout/components/Sidebar/components/Section/index.module.scss
