+++ packages/console/src/containers/AppLayout/components/Sidebar/icons/BarGraph.tsx
