+++ packages/console/src/containers/AppLayout/components/Sidebar/icons/Bolt.tsx
