+++ packages/console/src/containers/AppLayout/components/Sidebar/icons/Box.tsx
