+++ packages/console/src/containers/AppLayout/components/Sidebar/icons/Cloud.tsx
