+++ packages/console/src/containers/AppLayout/components/Sidebar/icons/Connection.tsx
