+++ packages/console/src/containers/AppLayout/components/Sidebar/icons/Document.tsx
