+++ packages/console/src/containers/AppLayout/components/Sidebar/icons/Gear.tsx
