+++ packages/console/src/containers/AppLayout/components/Sidebar/icons/UserProfile.tsx
