+++ packages/console/src/containers/AppLayout/components/Sidebar/icons/Web.tsx
