+++ packages/console/src/containers/AppLayout/components/Sidebar/index.module.scss
