+++ packages/console/src/containers/AppLayout/components/UserInfoSkeleton/index.module.scss
