+++ packages/console/src/containers/AppLayout/components/UserInfoSkeleton/index.tsx
