+++ packages/console/src/containers/AppLayout/types.ts
@@ -1,0 +1,5 @@
+import type { RefObject } from 'react';
+
+export type AppLayoutOutletContext = {
+  mainRef: RefObject<HTMLDivElement>;
+};
