+++ packages/console/src/pages/CloudPreview/index.tsx
@@ -1,0 +1,5 @@
+const CloudPreview = () => {
+  return <div>CloudPreview(TBD)</div>;
+};
+
+export default CloudPreview;
